SYCL interface: Move files that are needed for device passes to libkernel
