+++ include/hipSYCL/glue/omp/omp_kernel_launcher.hpp
@@ -37,13 +37,13 @@
 #include "hipSYCL/runtime/omp/omp_queue.hpp"
 #include "hipSYCL/sycl/backend/backend.hpp"
 #include "hipSYCL/sycl/exception.hpp"
-#include "hipSYCL/sycl/range.hpp"
-#include "hipSYCL/sycl/id.hpp"
-#include "hipSYCL/sycl/item.hpp"
-#include "hipSYCL/sycl/nd_item.hpp"
-#include "hipSYCL/sycl/sp_item.hpp"
-#include "hipSYCL/sycl/group.hpp"
-#include "hipSYCL/sycl/detail/local_memory_allocator.hpp"
+#include "hipSYCL/sycl/libkernel/range.hpp"
+#include "hipSYCL/sycl/libkernel/id.hpp"
+#include "hipSYCL/sycl/libkernel/item.hpp"
+#include "hipSYCL/sycl/libkernel/nd_item.hpp"
+#include "hipSYCL/sycl/libkernel/sp_item.hpp"
+#include "hipSYCL/sycl/libkernel/group.hpp"
+#include "hipSYCL/sycl/libkernel/detail/local_memory_allocator.hpp"
 
 #include "hipSYCL/runtime/device_id.hpp"
 #include "hipSYCL/runtime/kernel_launcher.hpp"
