+++ include/hipSYCL/sycl/device.hpp
@@ -37,8 +37,8 @@
 #include "info/info.hpp"
 #include "backend/backend.hpp"
 #include "exception.hpp"
-#include "id.hpp"
 #include "version.hpp"
+#include "hipSYCL/sycl/libkernel/id.hpp"
 
 #include "hipSYCL/runtime/device_id.hpp"
 #include "hipSYCL/runtime/application.hpp"
