+++ include/hipSYCL/sycl/handler.hpp
@@ -34,16 +34,16 @@
 
 #include "exception.hpp"
 #include "access.hpp"
-#include "accessor.hpp"
 #include "backend/backend.hpp"
 #include "device.hpp"
 #include "types.hpp"
-#include "id.hpp"
-#include "range.hpp"
-#include "nd_range.hpp"
-#include "item.hpp"
-#include "nd_item.hpp"
-#include "group.hpp"
+#include "libkernel/accessor.hpp"
+#include "libkernel/id.hpp"
+#include "libkernel/range.hpp"
+#include "libkernel/nd_range.hpp"
+#include "libkernel/item.hpp"
+#include "libkernel/nd_item.hpp"
+#include "libkernel/group.hpp"
 #include "detail/util.hpp"
 #include "detail/local_memory_allocator.hpp"
