+++ include/hipSYCL/sycl/libkernel/common_functions.hpp
