+++ include/hipSYCL/sycl/libkernel/detail/device_array.hpp
@@ -29,7 +29,7 @@
 #define HIPSYCL_DEVICE_ARRAY_HPP
 
 #include <type_traits>
-#include "../backend/backend.hpp"
+#include "hipSYCL/sycl/backend/backend.hpp"
 
 namespace hipsycl {
 namespace sycl {
