+++ include/hipSYCL/sycl/libkernel/detail/local_memory_allocator.hpp
@@ -28,8 +28,8 @@
 #ifndef HIPSYCL_LOCAL_MEM_ALLOCATOR_HPP
 #define HIPSYCL_LOCAL_MEM_ALLOCATOR_HPP
 
-#include "../backend/backend.hpp"
-#include "../exception.hpp"
+#include "hipSYCL/sycl/backend/backend.hpp"
+#include "hipSYCL/sycl/exception.hpp"
 
 #include <cstdlib>
 #include <array>
