+++ include/hipSYCL/sycl/libkernel/detail/mobile_shared_ptr.hpp
@@ -28,8 +28,8 @@
 #ifndef HIPSYCL_MOBILE_SHARED_PTR_HPP
 #define HIPSYCL_MOBILE_SHARED_PTR_HPP
 
-#include "../backend/backend.hpp"
-#include "../types.hpp"
+#include "hipSYCL/sycl/backend/backend.hpp"
+#include "hipSYCL/sycl/types.hpp"
 
 namespace hipsycl {
 namespace sycl {
