+++ include/hipSYCL/sycl/libkernel/detail/vec_common.hpp
