+++ include/hipSYCL/sycl/libkernel/detail/vec_swizzle.hpp
