+++ include/hipSYCL/sycl/libkernel/geometric_functions.hpp
