+++ include/hipSYCL/sycl/libkernel/item.hpp
@@ -28,10 +28,11 @@
 #ifndef HIPSYCL_ITEM_HPP
 #define HIPSYCL_ITEM_HPP
 
+#include "hipSYCL/sycl/backend/backend.hpp"
+
 #include "id.hpp"
 #include "range.hpp"
 #include "detail/thread_hierarchy.hpp"
-#include "backend/backend.hpp"
 
 #include <type_traits>
