+++ include/hipSYCL/sycl/libkernel/range.hpp
@@ -32,7 +32,7 @@
 #include <type_traits>
 #include <array>
 
-#include "backend/backend.hpp"
+#include "hipSYCL/sycl/backend/backend.hpp"
 #include "id.hpp"
 #include "detail/device_array.hpp"
 #include "hipSYCL/runtime/util.hpp"
