+++ include/hipSYCL/sycl/libkernel/sp_item.hpp
@@ -28,7 +28,7 @@
 #ifndef HIPSYCL_SP_ITEM_HPP
 #define HIPSYCL_SP_ITEM_HPP
 
-#include "backend/backend.hpp"
+#include "hipSYCL/sycl/backend/backend.hpp"
 #include "id.hpp"
 #include "range.hpp"
 #include "detail/thread_hierarchy.hpp"
