+++ include/hipSYCL/sycl/libkernel/stream.hpp
@@ -30,7 +30,7 @@
 
 #include <cstdio>
 
-#include "backend/backend.hpp"
+#include "hipSYCL/sycl/backend/backend.hpp"
 
 #include "id.hpp"
 #include "range.hpp"
